Allow compiling to WASM
By updating go-sockaddr to 1.0.6 and shuffling some files around
so that build tags would ignore irrelevant files for JS/WASM (such
as the plugincontainer stuff that doesn't even make sense when
compiling to JS/WASM).

Update to use hashicorp/cli instead of mitchellh/cli for JS/WASM support.
