+++ .github/workflows/go.yml
@@ -39,6 +39,9 @@ jobs:
     - name: Build
       run: cd ${{ matrix.module }} && go build ./...
 
+    - name: JS/WASM build
+      run: cd ${{ matrix.module }} && GOOS=js GOARCH=wasm go build -o test.wasm .
+
     - name: Install gVisor
       if: ${{ matrix.module == 'plugincontainer' }}
       run: |
@@ -92,7 +95,7 @@ jobs:
 
         export PATH="$HOME/bin:$PATH"
         systemctl --user restart docker
-    
+
     - name: Test
       run: cd ${{ matrix.module }} && go test ./...
