+++ listenerutil/listener.go
@@ -14,10 +14,10 @@ import (
 	osuser "os/user"
 	"strconv"
 
+	"github.com/hashicorp/cli"
 	"github.com/hashicorp/go-secure-stdlib/reloadutil"
 	"github.com/hashicorp/go-secure-stdlib/tlsutil"
 	"github.com/jefferai/isbadcipher"
-	"github.com/mitchellh/cli"
 )
 
 type Listener struct {
