+++ password/password_unix.go
@@ -1,8 +1,8 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-//go:build linux || darwin || freebsd || netbsd || openbsd || dragonfly
-// +build linux darwin freebsd netbsd openbsd dragonfly
+//go:build linux || darwin || freebsd || netbsd || openbsd || dragonfly || js
+// +build linux darwin freebsd netbsd openbsd dragonfly js
 
 package password
